% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highVolumeTimeSeries.R
\name{highVolumeTimeSeries}
\alias{highVolumeTimeSeries}
\title{Download entire time-series of High-Volume For-Hire records}
\usage{
highVolumeTimeSeries(from = 3, to = 6)
}
\arguments{
\item{from}{integer representing first month to scrape, i.e. 3 for march}

\item{to}{integer representing last month to scrape, i.e. 6 for june}
}
\value{
time-series of data-sets combined
}
\description{
{Scrape all months from \code{from} to \code{to} and combine into
one time-series.}
}
