% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highVolumeTimeSeries1.R
\name{highVolumeTimeSeries1}
\alias{highVolumeTimeSeries1}
\title{Download entire time-series of High-Volume For-Hire records}
\usage{
highVolumeTimeSeries1(month = 3)
}
\arguments{
\item{month}{integer representing first month to scrape, i.e. 3 for march}
}
\value{
time-series of data-sets combined
}
\description{
{Scrape all months from \code{from} to \code{to} and combine into
one time-series.}
}
